pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["Devansh3712 <devanshamity@gmail.com>"]
readme = "README.md"
packages = [{include = "drf_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
psycopg2 = "^2.9.5"
isort = "^5.12.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pre-commit = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
